Use erlang:nif_error/1 in nif wrapper
According to Scott Fritchie, the use of random:uniform/0
to occasionally call some native Erlang code
"was a really old style hack around a Dialyzer limitation"
and should be replaced with erlang:nif_error/1.

The random module was deprecated in OTP-19 and the compilation
of the ebloom module fails with warnings_as_errors.